requirements.txt
streamlit
torch
opencv-python
Pillow
numpy
streamlit==1.35.0
torch==2.2.2
opencv-python==4.9.0.80